+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Move timer to top of move list ([@Toscani](https://lichess.org/@/Toscani))
+* Automatically start timer ([@Toscani](https://lichess.org/@/Toscani))
 
 ## 1.0.0